requirements.txt
asgiref==3.7.2
boto3==1.28.5
botocore==1.31.5
dj-database-url==0.5.0
Django==3.2.20
django-allauth==0.41.0
django-ckeditor==6.6.1
django-countries==7.2.1
django-crispy-forms==1.14.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-storages==1.13.2
gunicorn==21.1.0
jmespath==1.0.1
oauthlib==3.2.2
Pillow==10.0.0
psycopg2==2.9.6
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
s3transfer==0.6.1
sqlparse==0.4.4
stripe==5.5.0
urllib3==1.26.16
